重构：统一模型结构和字段命名
- 标准化时间戳类型为int64
- 为所有模型添加TableName()方法
- 修复字段命名不一致问题：AuthorId→AuthorID，Url→URL，CountCount→UseCount
- 改进枚举类型：创建CommentStatus枚举类型
- 从模型中移除业务逻辑：时间格式化方法，缓存逻辑
- 修复JSON映射问题：Attachment.Title正确映射到"title"
- 更新所有测试文件以匹配重构后的模型
- 替换已弃用的函数调用

+++ model/category.go
@@ -1,11 +1,10 @@
 package model
 
 type Category struct {
-	ID   uint   `gorm:"primaryKey" json:"id"`
-	Text string `json:"text"`
-	Slug string `json:"slug"`
-
-	CreatedAt int `json:"createdAt,omitempty"  gorm:"autoCreateTime"`
+	ID        uint   `gorm:"primaryKey" json:"id"`
+	Text      string `json:"text"`
+	Slug      string `json:"slug"`
+	CreatedAt int64  `json:"createdAt,omitempty" gorm:"autoCreateTime"`
 }
 
 func (Category) TableName() string {
